refactor: move flattenObject to separate file
