+++ packages/consumption/src/modules/common/index.ts
@@ -1,1 +1,2 @@
+export * from "./flattenObject";
 export * from "./ValidationResult";
